Refactor: extracted admin stuff from main controller
